+++ src/client/components/SecondaryButton.jsx
@@ -3,8 +3,7 @@ import Button from '@govuk-react/button'
 import React from 'react'
 
 // TODO: Move to data-hub-components
-const SecondaryButton = (props) => (
+const SecondaryButton = props =>
   <Button buttonColour={GREY_3} buttonTextColour={TEXT_COLOUR} {...props} />
-)
 
 export default SecondaryButton
